+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Hide plot_number in plots view. [#634](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/634)
+
 ## [2.19.0](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/40) 2024-02-23
 
 ### Added